file-transfer-per-drive/requirements.txt
# Google Drive API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# File System Monitoring
watchdog>=3.0.0

# SSH/SCP Transfer
paramiko>=3.3.1
scp>=0.14.5

# Note: pathlib is part of Python standard library (Python 3.4+)
# No need to install separately

